requirements.txt
langchain
langchain-community
langchain-openai
langchain-core
streamlit
python-dotenv
deepeval
pypdf
python-docx
unstructured
librosa
soundfile
torch
transformers
pytesseract
pillow
qdrant-client
sentence-transformers
neo4j
tiktoken